+++ src/main/kotlin/com/example/payment/repository/PaymentUserRepositories.kt
@@ -19,4 +19,6 @@ interface OrderTransactionRepository : JpaRepository<OrderTransaction, Long> {
         order: Order,
         transactionType: TransactionType
     ): List<OrderTransaction>
+
+    fun findByTransactionId(originalTransactionId: String): OrderTransaction?
 }
