+++ src/main/kotlin/com/example/payment/service/payment/PaymentService.kt
@@ -1,7 +1,8 @@
-package com.example.payment.service
+package com.example.payment.service.payment
 
 import com.example.payment.exception.ErrorCode
 import com.example.payment.exception.PaymentException
+import com.example.payment.service.AccountService
 import org.springframework.stereotype.Service
 import java.time.LocalDateTime
 import javax.validation.constraints.Min
@@ -27,6 +28,7 @@ class PaymentService(
 
         return try {
             // 계좌에 금액 사용 요청(외부 시스템에 요청을 보내야함)
+            // payMethodTransactionId => accountSystem transactionId
             val payMethodTransactionId = this.accountService.useAccount(orderId)
 
             // 성공 : 거래를 성공적으로 저장
