+++ src/main/kotlin/com/example/payment/util/IdUtils.kt
@@ -6,4 +6,6 @@ fun generateOrderId() = "PO" + getUUID()
 
 fun generateTransactionId() = "PT" + getUUID()
 
+fun generateRefundTransactionId() = "PR" + getUUID()
+
 private fun getUUID() = UUID.randomUUID().toString().replace("-", "")
