+++ src/test/kotlin/com/example/payment/service/PaymentServiceTest.kt
@@ -2,6 +2,9 @@ package com.example.payment.service
 
 import com.example.payment.exception.ErrorCode.*
 import com.example.payment.exception.PaymentException
+import com.example.payment.service.payment.PayServiceRequest
+import com.example.payment.service.payment.PaymentService
+import com.example.payment.service.payment.PaymentStatusService
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
